+++ lib/cext/include/truffleruby/truffleruby-pre.h
@@ -71,6 +71,7 @@ extern VALUE (*rb_tr_wrap)(void *obj);
 extern VALUE (*rb_tr_longwrap)(long obj);
 extern void* (*rb_tr_id2sym)(ID obj);
 extern ID (*rb_tr_sym2id)(VALUE sym);
+extern void* (*rb_tr_force_native)(VALUE obj);
 
 #include <ruby/thread_native.h>