+++ src/main/java/org/truffleruby/core/format/convert/ToStringNode.java
@@ -12,18 +12,23 @@ package org.truffleruby.core.format.convert;
 import java.nio.charset.StandardCharsets;
 
 import com.oracle.truffle.api.library.CachedLibrary;
+import org.jcodings.specific.USASCIIEncoding;
+import org.jcodings.specific.UTF8Encoding;
 import org.truffleruby.core.array.RubyArray;
 import org.truffleruby.core.format.FormatNode;
 import org.truffleruby.core.format.exceptions.NoImplicitConversionException;
 import org.truffleruby.core.kernel.KernelNodes;
-import org.truffleruby.core.rope.RopeNodes;
+import org.truffleruby.core.klass.RubyClass;
+import org.truffleruby.core.rope.CodeRange;
+import org.truffleruby.core.rope.LazyIntRope;
+import org.truffleruby.core.rope.Rope;
+import org.truffleruby.core.rope.RopeConstants;
 import org.truffleruby.core.rope.RopeOperations;
 import org.truffleruby.language.Nil;
 import org.truffleruby.language.dispatch.DispatchNode;
 
 import com.oracle.truffle.api.CompilerDirectives;
 import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
-import com.oracle.truffle.api.dsl.Cached;
 import com.oracle.truffle.api.dsl.NodeChild;
 import com.oracle.truffle.api.dsl.Specialization;
 import com.oracle.truffle.api.frame.VirtualFrame;
@@ -38,6 +43,7 @@ public abstract class ToStringNode extends FormatNode {
     private final String conversionMethod;
     private final boolean inspectOnConversionFailure;
     private final Object valueOnNil;
+    protected final boolean specialClassBehaviour;
 
     @Child private DispatchNode toStrNode;
     @Child private DispatchNode toSNode;
@@ -48,10 +54,20 @@ public abstract class ToStringNode extends FormatNode {
             String conversionMethod,
             boolean inspectOnConversionFailure,
             Object valueOnNil) {
+        this(convertNumbersToStrings, conversionMethod, inspectOnConversionFailure, valueOnNil, false);
+    }
+
+    public ToStringNode(
+            boolean convertNumbersToStrings,
+            String conversionMethod,
+            boolean inspectOnConversionFailure,
+            Object valueOnNil,
+            boolean specialClassBehaviour) {
         this.convertNumbersToStrings = convertNumbersToStrings;
         this.conversionMethod = conversionMethod;
         this.inspectOnConversionFailure = inspectOnConversionFailure;
         this.valueOnNil = valueOnNil;
+        this.specialClassBehaviour = specialClassBehaviour;
     }
 
     public abstract Object executeToString(VirtualFrame frame, Object object);
@@ -63,43 +79,56 @@ public abstract class ToStringNode extends FormatNode {
 
     @TruffleBoundary
     @Specialization(guards = "convertNumbersToStrings")
-    protected byte[] toString(int value) {
-        return RopeOperations.encodeAsciiBytes(Integer.toString(value));
+    protected Rope toString(int value) {
+        return new LazyIntRope(value);
     }
 
     @TruffleBoundary
     @Specialization(guards = "convertNumbersToStrings")
-    protected byte[] toString(long value) {
-        return RopeOperations.encodeAsciiBytes(Long.toString(value));
+    protected Rope toString(long value) {
+        return RopeOperations.encodeAscii(Long.toString(value), USASCIIEncoding.INSTANCE);
     }
 
     @TruffleBoundary
     @Specialization(guards = "convertNumbersToStrings")
-    protected byte[] toString(double value) {
-        return RopeOperations.encodeAsciiBytes(Double.toString(value));
+    protected Rope toString(double value) {
+        return RopeOperations.encodeAscii(Double.toString(value), USASCIIEncoding.INSTANCE);
+    }
+
+    @TruffleBoundary
+    @Specialization(guards = "specialClassBehaviour")
+    protected Rope toStringSpecialClass(RubyClass rubyClass,
+            @CachedLibrary(limit = "2") RubyStringLibrary libString) {
+        if (rubyClass == getContext().getCoreLibrary().trueClass) {
+            return RopeConstants.TRUE;
+        } else if (rubyClass == getContext().getCoreLibrary().falseClass) {
+            return RopeConstants.FALSE;
+        } else if (rubyClass == getContext().getCoreLibrary().nilClass) {
+            return RopeConstants.NIL;
+        } else {
+            return toString(rubyClass, libString);
+        }
     }
 
     @Specialization(guards = "libString.isRubyString(string)")
-    protected byte[] toStringString(Object string,
+    protected Rope toStringString(Object string,
             @CachedLibrary(limit = "2") RubyStringLibrary libValue,
-            @CachedLibrary(limit = "2") RubyStringLibrary libString,
-            @Cached RopeNodes.BytesNode bytesNode) {
+            @CachedLibrary(limit = "2") RubyStringLibrary libString) {
         if ("inspect".equals(conversionMethod)) {
             final Object value = getToStrNode().call(string, conversionMethod);
 
             if (libValue.isRubyString(value)) {
-                return bytesNode.execute(libValue.getRope(value));
+                return libValue.getRope(value);
             } else {
                 throw new NoImplicitConversionException(string, "String");
             }
         }
-        return bytesNode.execute(libString.getRope(string));
+        return libString.getRope(string);
     }
 
     @Specialization
-    protected byte[] toString(RubyArray array,
-            @CachedLibrary(limit = "2") RubyStringLibrary libString,
-            @Cached RopeNodes.BytesNode bytesNode) {
+    protected Rope toString(RubyArray array,
+            @CachedLibrary(limit = "2") RubyStringLibrary libString) {
         if (toSNode == null) {
             CompilerDirectives.transferToInterpreterAndInvalidate();
             toSNode = insert(DispatchNode.create(PRIVATE_RETURN_MISSING));
@@ -108,7 +137,7 @@ public abstract class ToStringNode extends FormatNode {
         final Object value = toSNode.call(array, "to_s");
 
         if (libString.isRubyString(value)) {
-            return bytesNode.execute(libString.getRope(value));
+            return libString.getRope(value);
         } else {
             throw new NoImplicitConversionException(array, "String");
         }
@@ -116,13 +145,12 @@ public abstract class ToStringNode extends FormatNode {
 
     @Specialization(
             guards = { "isNotRubyString(object)", "!isRubyArray(object)", "!isForeignObject(object)" })
-    protected byte[] toString(Object object,
-            @CachedLibrary(limit = "2") RubyStringLibrary libString,
-            @Cached RopeNodes.BytesNode bytesNode) {
+    protected Rope toString(Object object,
+            @CachedLibrary(limit = "2") RubyStringLibrary libString) {
         final Object value = getToStrNode().call(object, conversionMethod);
 
         if (libString.isRubyString(value)) {
-            return bytesNode.execute(libString.getRope(value));
+            return libString.getRope(value);
         }
 
         if (inspectOnConversionFailure) {
@@ -131,7 +159,7 @@ public abstract class ToStringNode extends FormatNode {
                 inspectNode = insert(KernelNodes.ToSNode.create());
             }
 
-            return bytesNode.execute(inspectNode.executeToS(object).rope);
+            return inspectNode.executeToS(object).rope;
         } else {
             throw new NoImplicitConversionException(object, "String");
         }
@@ -139,8 +167,11 @@ public abstract class ToStringNode extends FormatNode {
 
     @TruffleBoundary
     @Specialization(guards = "isForeignObject(object)")
-    protected byte[] toStringForeign(Object object) {
-        return object.toString().getBytes(StandardCharsets.UTF_8);
+    protected Rope toStringForeign(Object object) {
+        return RopeOperations.create(
+                object.toString().getBytes(StandardCharsets.UTF_8),
+                UTF8Encoding.INSTANCE,
+                CodeRange.CR_UNKNOWN);
     }
 
     private DispatchNode getToStrNode() {