+++ src/main/java/org/truffleruby/core/format/read/array/ReadStringNode.java
@@ -10,8 +10,6 @@
 package org.truffleruby.core.format.read.array;
 
 import org.truffleruby.core.array.ArrayGuards;
-import org.truffleruby.core.format.TruncateStringNode;
-import org.truffleruby.core.format.TruncateStringNodeGen;
 import org.truffleruby.core.array.library.ArrayStoreLibrary;
 import org.truffleruby.core.format.FormatNode;
 import org.truffleruby.core.format.LiteralFormatNode;
@@ -35,22 +33,29 @@ public abstract class ReadStringNode extends FormatNode {
     private final String conversionMethod;
     private final boolean inspectOnConversionFailure;
     private final Object valueOnNil;
-    private final Integer precision;
+    private final boolean specialClassBehaviour;
 
     @Child private ToStringNode toStringNode;
-    @Child private TruncateStringNode truncateStringNode;
+
+    public ReadStringNode(
+            boolean convertNumbersToStrings,
+            String conversionMethod,
+            boolean inspectOnConversionFailure,
+            Object valueOnNil) {
+        this(convertNumbersToStrings, conversionMethod, inspectOnConversionFailure, valueOnNil, false);
+    }
 
     public ReadStringNode(
             boolean convertNumbersToStrings,
             String conversionMethod,
             boolean inspectOnConversionFailure,
             Object valueOnNil,
-            Integer precision) {
+            boolean specialClassBehaviour) {
         this.convertNumbersToStrings = convertNumbersToStrings;
         this.conversionMethod = conversionMethod;
         this.inspectOnConversionFailure = inspectOnConversionFailure;
         this.valueOnNil = valueOnNil;
-        this.precision = precision;
+        this.specialClassBehaviour = specialClassBehaviour;
     }
 
     @Specialization(limit = "storageStrategyLimit()")
@@ -67,15 +72,8 @@ public abstract class ReadStringNode extends FormatNode {
                     conversionMethod,
                     inspectOnConversionFailure,
                     valueOnNil,
+                    specialClassBehaviour,
                     WriteByteNodeGen.create(new LiteralFormatNode((byte) 0))));
-        }
-
-        if (this.precision != null) {
-            if (truncateStringNode == null) {
-                CompilerDirectives.transferToInterpreterAndInvalidate();
-                this.truncateStringNode = insert(TruncateStringNodeGen.create(precision, null));
-            }
-            value = truncateStringNode.execute(value);
         }
 
         return toStringNode.executeToString(frame, value);